% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_srkw_data.R
\name{load_srkw_data}
\alias{load_srkw_data}
\title{Load Southern Resident killer whale population data}
\usage{
load_srkw_data()
}
\value{
A data frame including SRKW population data
}
\description{
Loads the data set containing yearly census counts for
Southern Resident killer whales
}
