% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srkw_shiny_app.R
\name{srkw_shiny_app}
\alias{srkw_shiny_app}
\title{Southern Resident killer whale population}
\usage{
srkw_shiny_app()
}
\value{
A shiny app launched in the browser or Viewer pane.
}
\description{
Gives an interactive shiny application for exploring the endangered
Southern Resident killer whale population trends, pod counts, births and
mortality, over 50 years.
}
\details{
The app uses the SRKW data set included in the package and provides
interactive visualizations of population trends, births and mortality across
years (1975-2024). Data provided by the Center for Whale Research (CWR) and
National Oceanic and Atmospheric Administration (NOAA).
}
\examples{
\dontrun{
# Launch the interactive Shiny app
srkw_shiny_app()
}

}
\references{
Center for Whale Research. Southern Resident Killer Whale
population data.\url{https://www.whaleresearch.com/orca-population/}

National Oceanic and Atmospheric Administration (NOAA).
\url{https://www.noaa.gov/}
}
